refactor: replace modal editors with in-canvas nodes
